Address PR feedback
Remove individual directories in BlockChain::open on WithExisting::Kill
and save genesis block details RLP in local variable before inserting
into extras db.

Untabify BlockDetails and pass ctor args by const&

Cleanup web3.eth.getBlock* calls

Make BlockDetails::blockSizeBytes a size_t instead of unsigned

Minor formatting/logging changes (e.g. untabify BlockDetails files)

+++ libethereum/BlockDetails.cpp
@@ -12,12 +12,12 @@ using namespace dev::eth;
 
 BlockDetails::BlockDetails(RLP const& _r)
 {
-	number = _r[0].toInt<unsigned>();
-	totalDifficulty = _r[1].toInt<u256>();
-	parentHash = _r[2].toHash<h256>();
-	childHashes = _r[3].toVector<h256>();
-	size = _r.size();
-	blockSizeBytes = _r[4].toInt<unsigned>();
+    number = _r[0].toInt<unsigned>();
+    totalDifficulty = _r[1].toInt<u256>();
+    parentHash = _r[2].toHash<h256>();
+    childHashes = _r[3].toVector<h256>();
+    size = _r.size();
+    blockSizeBytes = _r[4].toInt<size_t>();
 }
 
 bytes BlockDetails::rlp() const
